requirements.txt
discord.py==2.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.12.14
requests==2.32.4
beautifulsoup4==4.12.2
aiosqlite==0.19.0  # Async SQLite for AI database

# Terminal bot dependencies
colorama==0.4.6  # Cross-platform colored terminal output
textual==0.47.1  # Modern TUI framework (optional)

# Development dependencies (optional)

watchdog==3.0.0  # For auto-restart on file changes

# AI Training dependencies (optional - only needed for training custom models)
# Uncomment these if you want to train your own models:
# transformers>=4.35.0  # Hugging Face transformers
# datasets>=2.14.0  # Dataset handling
# accelerate>=0.24.0  # Training acceleration
# peft>=0.6.0  # Parameter-efficient fine-tuning (LoRA)
# torch>=2.1.0  # PyTorch (CPU version)
# For GPU support, install torch separately:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118